requirements.txt
adal==1.2.2
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.4
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-sendgrid-v5==0.8.0
djangorestframework==3.10.2
future==0.17.1
gunicorn==19.9.0
idna==2.8
Pillow==6.1.0
pprint==0.1
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-http-client==3.1.0
pytz==2019.2
requests==2.22.0
sendgrid==6.0.5
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
